Established Edge Cases
The first edge case would be non-integer inputs such as 'ab'. The program will parse those inputs as 'NaN'. An exception would be thrown if one of them is NaN to tell the user to only input integer values.

The second edge case would be a zero input. If we use the current algorithm, it will see that the input would be zero; thus ending the program. Hence, I implemented an edge case at the beginning of the algorithm to check if it’s zero.

The third edge case would be a negative input. Integers are whole numbers that can be negative or positive. The algorithm assumes all inputs are positive integers. If a negative input is processed, it will not be able to map the phonetic value to the negative number. I implemented an edge case to handle any negative integer value.